# internal/service/sail/docs/swagger.yaml
definitions:
  model.AddConfig:
    properties:
      content:
        type: string
      is_encrypt:
        type: boolean
      is_link_public:
        type: boolean
      is_public:
        type: boolean
      name:
        type: string
      namespace_id:
        type: integer
      project_group_id:
        type: integer
      project_id:
        description: 公共配置可以不传projectID
        type: integer
      public_config_id:
        type: integer
      type:
        type: string
    type: object
  model.AddNamespace:
    properties:
      name:
        type: string
      project_group_id:
        type: integer
      real_time:
        description: 是否灰度
        type: boolean
      secret:
        description: 是否加密
        type: boolean
    type: object
...
  model.ConfigCopy:
    properties:
      config_id:
        type: integer
      op:
        description: 1 转为副本 2关联公共配置
        type: integer
    type: object
  model.ConfigHistoryList:
    properties:
      config_id:
        type: integer
      create_by:
        type: integer
      create_by_name:
        type: string
      create_time:
        type: integer
      op_type:
        type: integer
      op_type_str:
        type: string
      reversion:
        type: integer
    type: object
  model.ConfigInfo:
    properties:
      config_id:
        type: integer
      config_key:
        type: string
      content:
        type: string
      is_encrypt:
        type: boolean
      is_link_public:
        type: boolean
      is_public:
        type: boolean
      name:
        type: string
      type:
        type: string
    type: object
  model.ConfigNode:
    properties:
      config_id:
        type: integer
      name:
        type: string
      type:
        type: string
    type: object
  model.EditConfig:
    properties:
      config_id:
        type: integer
      content:
        type: string
    type: object
  model.EditNamespace:
    properties:
      name:
        type: string
      namespace_id:
        type: integer
      real_time:
        type: boolean
    type: object
...
  model.NamespaceList:
    properties:
      create_by:
        type: integer
      create_by_name:
        type: string
      create_time:
        type: integer
      name:
        type: string
      namespace_id:
        type: integer
      project_group_id:
        type: integer
      project_group_name:
        type: string
      real_time:
        description: 是否灰度
        type: boolean
      secret_key:
        type: string
    type: object
...
  model.ProjectTree:
    properties:
      can_secret:
        description: 是否能加密
        type: boolean
      name:
        type: string
      namespace_id:
        type: integer
      nodes:
        items:
          $ref: '#/definitions/model.ConfigNode'
        type: array
      real_time:
        description: 是否需发布
        type: boolean
    type: object
...
  model.RollbackConfig:
    properties:
      config_id:
        type: integer
      reversion:
        type: integer
    type: object
...
  page.Page:
    properties:
      count:
        type: integer
      list: {}
    type: object
  response.JsonResponse:
    properties:
      code:
        description: 业务码
        type: integer
      data: {}
      message:
        description: 描述信息
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: FengYu
    url: https://hyy-yu.space
  description: 配置中心接口设计文档
  title: 配置中心
  version: "1.0"
paths:
  /v1/config/add:
    post:
      parameters:
      - description: data
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/model.AddConfig'
      responses:
        "200":
          description: data=ok
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  type: string
              type: object
      summary: 添加配置
      tags:
      - 配置管理
  /v1/config/copy:
    post:
      parameters:
      - description: data
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/model.ConfigCopy'
      responses:
        "200":
          description: data=ok
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  type: string
              type: object
      summary: 副本配置
      tags:
      - 配置管理
  /v1/config/del:
    post:
      parameters:
      - description: ConfigId
        in: body
        name: config_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          description: data=ok
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  type: string
              type: object
      summary: 删除配置
      tags:
      - 配置管理
  /v1/config/edit:
    post:
      parameters:
      - description: data
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/model.EditConfig'
      responses:
        "200":
          description: data=ok
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  type: string
              type: object
      summary: 编辑配置
      tags:
      - 配置管理
  /v1/config/history:
    get:
      parameters:
      - description: 配置ID
        in: query
        name: config_id
        type: integer
      responses:
        "200":
          description: data
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.ConfigHistoryList'
                  type: array
              type: object
      summary: 配置历史
      tags:
      - 配置管理
  /v1/config/history_info:
    get:
      parameters:
      - description: 配置ID
        in: query
        name: config_id
        type: integer
      - description: reversion
        in: query
        name: reversion
        type: integer
      responses:
        "200":
          description: data
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  type: string
              type: object
      summary: 配置历史详情
      tags:
      - 配置管理
  /v1/config/info:
    get:
      parameters:
      - description: 配置ID
        in: query
        name: config_id
        type: integer
      responses:
        "200":
          description: data
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.ConfigInfo'
              type: object
      summary: 配置详情
      tags:
      - 配置管理
  /v1/config/rollback:
    post:
      parameters:
      - description: data
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/model.RollbackConfig'
      responses:
        "200":
          description: data=ok
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  type: string
              type: object
      summary: 回滚配置
      tags:
      - 配置管理
  /v1/config/tree:
    get:
      parameters:
      - description: 配置ID
        in: query
        name: project_id
        type: integer
      - description: 配置组ID
        in: query
        name: project_group_id
        type: integer
      responses:
        "200":
          description: data
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.ProjectTree'
                  type: array
              type: object
      summary: 配置树
      tags:
      - 配置管理
...
  /v1/namespace/add:
    post:
      parameters:
      - description: data
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/model.AddNamespace'
      responses:
        "200":
          description: data=ok
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  type: string
              type: object
      summary: 添加命名空间
      tags:
      - 命名空间管理
  /v1/namespace/del:
    post:
      parameters:
      - description: NamespaceId
        in: body
        name: namespace_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          description: data=ok
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  type: string
              type: object
      summary: 删除命名空间
      tags:
      - 命名空间管理
  /v1/namespace/edit:
    post:
      parameters:
      - description: data
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/model.EditNamespace'
      responses:
        "200":
          description: data=ok
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  type: string
              type: object
      summary: 编辑命名空间
      tags:
      - 命名空间管理
  /v1/namespace/list:
    get:
      parameters:
      - default: 1
        description: 页号
        in: query
        name: page_index
        type: integer
      - default: 10
        description: 页长
        in: query
        name: page_size
        type: integer
      - description: 排序字段
        in: query
        name: sort
        type: string
      - description: ID
        in: query
        name: project_group_id
        required: true
        type: integer
      - description: 命名空间ID
        in: query
        name: namespace_id
        type: integer
      - description: 命名空间名称
        in: query
        name: namespace_name
        type: string
      responses:
        "200":
          description: data
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/page.Page'
                  - properties:
                      List:
                        $ref: '#/definitions/model.NamespaceList'
                    type: object
              type: object
      summary: 命名空间列表
      tags:
      - 命名空间管理
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"